+++ Yeti/Controllers/FeedVC/FeedVC+Actions.m
@@ -281,7 +281,7 @@
             sender.enabled = YES;
         });
         
-        [[UNUserNotificationCenter currentNotificationCenter] requestAuthorizationWithOptions:UNAuthorizationOptionBadge|UNAuthorizationOptionAlert|UNAuthorizationOptionSound completionHandler:^(BOOL granted, NSError * _Nullable error) {
+        [self.mainCoordinator registerForNotifications:^(BOOL granted, NSError * _Nullable error) {
             
             if (error) {
                 NSLog(@"Error authorizing for push notifications: %@",error);